requirements.txt
Django~=5.0.4
django-resized
djoser
djangorestframework~=3.15.1
face-recognition
face-recognition-models
opencv-python~=4.9.0.80
gunicorn
numpy~=1.26.4
django-filter~=24.2